+++ src/test/kotlin/de/magynhard/crystal/parser/CrystalParserTest.kt
@@ -267,6 +267,10 @@ class CrystalParserTest : ParsingTestCase("", "cr", CrystalParserDefinition()) {
         doTest(true)
     }
 
+    fun testMacroMultiAssignTarget() {
+        doTest(true)
+    }
+
     fun testMacroInterpolatedCallee() {
         doTest(true)
     }